press releases: use generic matching rule
This way instead of having to explicitly add releases to the urls.py
file, we just add the relase to the static_templates/press_releases
directory and the right thing happens.

Getting rid of the current "feature" where /pressrelease redirects to
the current press release.  It was broken, looks like it's unused.

Limitation: release file name must match slug, but with underscores and
all lower case.

Why bother? I wanted to do something small to learn the edX toolchain.