Adding import of overrides file
Theme overrides should be placed in `base/_overrides.scss` so that
they'll be picked up by the SASS compiler.

A cleaner way to do this would be to set a THEME variable (or something
like that) in the settings, and then run `application.scss` through
the Mako template engine so that we could swap in the appropriate
theme file(s). However, I don't see a particularly clean way to do
that right now.