Initial commit of Stanford header/footer
Hacking in Stanford's header and footer. It's not very pretty:
requires some ugly changes to the main templates and to the settings,
but it does the trick.

The main CSS file has been edited from the source at
identity.stanford.edu to avoid conflicting ids and class names on
elements. Still need to edit the IE fixes.