Initial commit for feature BLD-391: Add arrows to the interface for start/end times.
If end time is specified (which means that there is also a start time), we
add an extra DOM element to the slider that will show the are in time between
start time and end time. Unfortunately, this visual aid will become available
ony after start of playback. This is because until then we don't know the length
of the video.